Let's go le calendrier bien cheum

+++ db/schema.rb
@@ -16,8 +16,8 @@ ActiveRecord::Schema.define(version: 2022_05_24_143321) do
   enable_extension "plpgsql"
 
   create_table "bookings", force: :cascade do |t|
-    t.bigint "user_id", null: false
-    t.bigint "superpower_id", null: false
+    t.integer "user_id", null: false
+    t.integer "superpower_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.date "starts_at"
@@ -30,7 +30,7 @@ ActiveRecord::Schema.define(version: 2022_05_24_143321) do
     t.string "name"
     t.text "description"
     t.integer "price"
-    t.bigint "user_id", null: false
+    t.integer "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["user_id"], name: "index_superpowers_on_user_id"
